+++ packages/cad-simple-viewer/src/view/AcTrView2d.ts
@@ -30,6 +30,7 @@ import {
   AcEdCalculateSizeCallback,
   AcEdConditionWaiter,
   AcEdCorsorType,
+  AcEdSpatialQueryResultItemEx,
   AcEdViewMode,
   eventBus
 } from '../editor'
@@ -438,7 +439,7 @@ export class AcTrView2d extends AcEdBaseView {
    */
   pick(point?: AcGePoint2dLike, hitRadius?: number) {
     if (point == null) point = this.curPos
-    const results: AcDbObjectId[] = []
+    const results: AcEdSpatialQueryResultItemEx[] = []
     const activeLayout = this._scene.activeLayout
     if (activeLayout) {
       const activeLayoutView = this.activeLayoutView
@@ -453,7 +454,7 @@ export class AcTrView2d extends AcEdBaseView {
       firstQueryResults.forEach(item => {
         const objectId = item.id
         if (activeLayout.isIntersectWith(objectId, raycaster)) {
-          results.push(objectId)
+          results.push(item)
         }
       })
     }
@@ -473,7 +474,7 @@ export class AcTrView2d extends AcEdBaseView {
   select(point?: AcGePoint2dLike) {
     const idsAdded: Array<AcDbObjectId> = []
     const results = this.pick(point)
-    results.forEach(id => idsAdded.push(id))
+    results.forEach(item => idsAdded.push(item.id))
     if (idsAdded.length > 0) this.selectionSet.add(idsAdded)
   }
 
@@ -664,11 +665,7 @@ export class AcTrView2d extends AcEdBaseView {
   }
 
   protected createScene() {
-    const scene = new AcTrScene()
-    scene.layouts.forEach(layout => {
-      layout.setSnapObject(this.renderer.createObject())
-    })
-    return scene
+    return new AcTrScene()
   }
 
   private createStats(show?: boolean) {
@@ -850,6 +847,8 @@ export class AcTrView2d extends AcEdBaseView {
       this._scene.addEntity(entity, true)
       entity.dispose()
     })
+    group.dispose()
+
     this._isDirty = true
   }