+++ packages/cad-viewer/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mlightcad/cad-viewer
 
+## 1.4.0
+
+### Minor Changes
+
+- feat: support osnap for insert entity
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/cad-simple-viewer@1.4.0
+
 ## 1.3.4
 
 ### Patch Changes