+++ packages/svg-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/svg-renderer
 
+## 1.4.0
+
+### Minor Changes
+
+- feat: support osnap for insert entity
+
 ## 1.3.2
 
 ### Patch Changes